fix: remove @flue/runtime/internal from adapter, keep root configs clean
The adapter no longer imports from @flue/runtime/internal — it only
imports public types from @flue/runtime. The user owns the Flue
runtime lifecycle in their run function and wires the adapter's event
handler via ctx.subscribeEvent(). This removes the need for bundler
module resolution in the package, the just-bash dependency, and all
root tsconfig workarounds.

The demo app uses @flue/runtime/internal since it needs to construct
the FlueContext for testing. Only apps/demo-flue is excluded from the
root typecheck (same pattern as packages/docs).
