+++ package.json
@@ -18,7 +18,7 @@
     "lint:fix": "biome lint --write .",
     "prepare": "simple-git-hooks",
     "release:check": "node ./scripts/check-release-config.mjs",
-    "typecheck": "tsc --noEmit -p tsconfig.typecheck.json",
+    "typecheck": "tsc --noEmit",
     "test": "dotenv -e .env -e .env.local -- vitest run packages apps scripts --config=./vitest.config.ts --reporter=./packages/vitest-evals/src/reporter.ts",
     "test:watch": "dotenv -e .env -e .env.local -- vitest packages apps scripts --config=./vitest.config.ts --reporter=./packages/vitest-evals/src/reporter.ts",
     "test:ci": "dotenv -e .env -e .env.local -- vitest run packages apps scripts --config=./vitest.config.ts --coverage --reporter=./packages/vitest-evals/src/reporter.ts --reporter=junit --outputFile=tests.junit.xml"
