+++ packages/harness-flue/tsconfig.json
@@ -1,8 +1,4 @@
 {
   "extends": "../../tsconfig.base.json",
-  "compilerOptions": {
-    "module": "esnext",
-    "moduleResolution": "bundler"
-  },
   "include": ["src/**/*.ts"]
 }
