+++ tsconfig.json
@@ -6,7 +6,6 @@
     "**/coverage/**",
     "node_modules",
     "packages/docs/src/content.config.ts",
-    "packages/harness-flue/**",
     "apps/demo-flue/**"
   ]
 }
